Add setting for JS touch events API
Currently we enable the touch events API by default if
a touch screen is detected by the system.
Some websites use this api to decide if they are on a
mobile device or desktop and base their design on it.
This can cause unwanted results on touch screen laptops
or other setups that emulate a fake touch device.

Add option to explicitly set the touch events API.
Keep old behavior if the option is not used.

[ChangeLog][Settings] Added JSTouchEventsEnabled setting

Task-number: QTBUG-115764
Change-Id: Ica860f54557ee21f00c63a3d30f9074a0278bc8c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>